feat: deleteNetwork

+++ src/controllers/network-controller.ts
@@ -64,11 +64,21 @@ export async function deleteNetwork(req: AuthenticatedRequest, res: Response){
     const networkId = req.params.id;
     const {userId} = req;
 
+    if(!networkId){
+        return res.status(400).send("id not sended!")
+    }
+
     try{
         const result = await networkService.deleteNetwork(Number(networkId), userId)
-        return res.send(result)
+        return res.sendStatus(204)
     } catch(err){
         console.log(err)
+        if(err.name==="NotFoundError"){
+            return res.status(404).send(err.message)
+        }
+        if(err.name==="forbiddenError"){
+            return res.status(403).send(err.message)
+        }
         return res.status(500).send(httpStatus["500_MESSAGE"])
     }
 }
